requirements.txt
selenium
anticaptchaofficial
Pillow
argparse
DateTime
regex
pyinstaller